(27/04) Correção de autorização quebrada de objeto no back-end
