+++ front-end/src/pages/UserList.css
@@ -1,0 +1,9 @@
+table {
+  margin: 0 auto;
+  border-collapse: collapse;
+}
+
+td, th {
+  border: 1px solid #888;
+  padding: 6px;
+}
